feat: 음식 API 모델 변경

+++ src/main/java/com/hbd/mommy/domain/wishlist/model/dto/request/RequestAddWishlist.java
@@ -8,5 +8,5 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class RequestAddWishlist {
 	@NotNull
-	private int foodId;
+	private Long foodId;
 }
